third_party/blink/web_tests/platform/generic/external/wpt/css/css-sizing/contain-intrinsic-size/contain-intrinsic-size-028-expected.txt

This is a testharness.js-based test.
PASS .test 1
PASS .test 2
PASS .test 3
PASS .test 4
PASS .test 5
PASS .test 6
PASS .test 7
PASS .test 8
FAIL .test 9 assert_equals: 
<img class="test cis-none" src="/css/support/60x60-green.png" data-expected-client-width="60" data-expected-client-height="60">
clientWidth expected 60 but got 0
FAIL .test 10 assert_equals: 
<img class="test cis-height" src="/css/support/60x60-green.png" data-expected-client-width="60" data-expected-client-height="60">
clientWidth expected 60 but got 0
PASS .test 11
PASS .test 12
FAIL .test 13 assert_equals: 
<img class="test cis-none vertical" src="/css/support/60x60-green.png" data-expected-client-width="60" data-expected-client-height="60">
clientHeight expected 60 but got 0
PASS .test 14
FAIL .test 15 assert_equals: 
<img class="test cis-width vertical" src="/css/support/60x60-green.png" data-expected-client-width="60" data-expected-client-height="60">
clientHeight expected 60 but got 0
PASS .test 16
FAIL .test 17 assert_equals: 
<svg class="test cis-none" data-expected-client-width="300" data-expected-client-height="150"></svg>
clientWidth expected 300 but got 0
FAIL .test 18 assert_equals: 
<svg class="test cis-height" data-expected-client-width="300" data-expected-client-height="150"></svg>
clientWidth expected 300 but got 0
PASS .test 19
PASS .test 20
FAIL .test 21 assert_equals: 
<svg class="test cis-none vertical" data-expected-client-width="300" data-expected-client-height="150"></svg>
clientHeight expected 150 but got 0
PASS .test 22
FAIL .test 23 assert_equals: 
<svg class="test cis-width vertical" data-expected-client-width="300" data-expected-client-height="150"></svg>
clientHeight expected 150 but got 0
PASS .test 24
Harness: the test ran to completion.
